refactor: move auth hook to entities
