+++ src/app/router/protected-route.tsx
@@ -1,7 +1,7 @@
 import { Navigate, useLocation } from 'react-router-dom';
 
+import { useAuth } from '@/entities/session';
 import { ROUTES } from '@/shared/config/routes';
-import { useAuth } from '@/shared/hooks/use-auth';
 import { Loader } from '@/shared/ui';
 
 interface ProtectedRouteProps {
