+++ src/entities/session/index.ts
@@ -1,0 +1,2 @@
+export { useAuth } from './hooks/use-auth';
+export { AuthContext } from './model/auth-context';
