+++ src/pages/profile/hooks/use-user-profile.ts
@@ -1,4 +1,4 @@
-import { useAuth } from '@/shared/hooks';
+import { useAuth } from '@/entities/session';
 
 export const useUserProfile = () => {
   const { user } = useAuth();
