+++ src/shared/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './use-auth';
 export * from './use-theme';
